Split legacy logic into multiple files
