+++ data/legacy.go
@@ -1,22 +1,13 @@
 package data
 
 import (
-	"fmt"
-	"net/url"
-	"strings"
-	"time"
-
 	"github.com/pokt-network/poktroll/pkg/polylog"
-	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/buildwithgrove/path/observation"
-	protocolobservation "github.com/buildwithgrove/path/observation/protocol"
-	qosobservation "github.com/buildwithgrove/path/observation/qos"
 )
 
 // TODO_MVP(@adshmh): Remove once the data pipeline has been updated.
 //
-
 // legacyRecord contains all the fields required by the legacy data pipeline.
 type legacyRecord struct {
 	RequestID              string  `json:"request_id"`            // Request's ID
@@ -46,7 +37,6 @@ type legacyRecord struct {
 
 	// internal fields used for tracking protocol and QoS data.
 	endpointTripTime float64 // endpoint response timestamp - endpoint query time, in seconds.
-
 }
 
 // converts a data record to legacy format, for compatibility with the existing data pipeline.
@@ -94,237 +84,4 @@ func buildLegacyDataRecord(
 	legacyRecord.PortalTripTime = legacyRecord.RequestRoundTripTime - legacyRecord.endpointTripTime
 
 	return legacyRecord
-}
-
-func setLegacyFieldsFromGatewayAuthData(
-	legacyRecord *legacyRecord,
-	authObservations *observation.RequestAuth,
-) *legacyRecord {
-	legacyRecord.RequestID = authObservations.RequestId
-	legacyRecord.Region = authObservations.Region
-
-	legacyRecord.PortalAccountID = authObservations.PortalAccountId
-	legacyRecord.PortalAppID = authObservations.PortalApplicationId
-
-	return legacyRecord
-}
-
-func setLegacyFieldsFromGatewayObservations(
-	logger polylog.Logger,
-	legacyRecord *legacyRecord,
-	observations *observation.GatewayObservations,
-) *legacyRecord {
-	legacyRecord = setLegacyFieldsFromGatewayAuthData(legacyRecord, observations.GetRequestAuth())
-
-	// Track organic (i.e. from the user) and synthetic (i.e. from the endpoint hydrator) requests.
-	legacyRecord.RelayType = observations.RequestType.String()
-
-	// Update request reception and completion timestamps.
-	legacyRecord.RequestStartTimestamp = formatTimestampPbForBigQueryJSON(observations.ReceivedTime)
-	legacyRecord.RequestReturnTimestamp = formatTimestampPbForBigQueryJSON(observations.CompletedTime)
-
-	// BigQuery DATE type, format: "2025-04-11"
-	legacyRecord.RequestDate = observations.ReceivedTime.AsTime().Format("2006-01-02")
-
-	// BigQuery TIMESTAMP type, format: "2025-04-11T14:30:00.000Z"
-	legacyRecord.RequestTimestamp = formatTimestampPbForBigQueryJSON(observations.ReceivedTime)
-
-	// Request processing time, in seconds.
-	legacyRecord.RequestRoundTripTime = observations.CompletedTime.AsTime().Sub(observations.ReceivedTime.AsTime()).Seconds()
-
-	return legacyRecord
-}
-
-// TODO_MVP(@adshmh): handle QoS observations for:
-// - Solana
-// - CometBFT
-func setLegacyFieldsFromQoSObservations(
-	logger polylog.Logger,
-	legacyRecord *legacyRecord,
-	observations *qosobservation.Observations,
-) *legacyRecord {
-	if evmObservations := observations.GetEvm(); evmObservations != nil {
-		return setLegacyFieldsFromQoSEVMObservations(logger, legacyRecord, evmObservations)
-	}
-
-	return legacyRecord
-}
-
-const qosEVMErrorTypeStr = "QOS_EVM"
-
-func setLegacyFieldsFromQoSEVMObservations(
-	logger polylog.Logger,
-	legacyRecord *legacyRecord,
-	observations *qosobservation.EVMRequestObservations,
-) *legacyRecord {
-	// In bytes: the length of the request: float64 type is for compatibility with the legacy data pipeline.
-	legacyRecord.RequestDataSize = float64(observations.RequestPayloadLength)
-
-	evmInterpreter := &qosobservation.EVMObservationInterpreter{
-		Observations: observations,
-	}
-
-	// Extract the JSONRPC request's method.
-	jsonrpcRequestMethod, _ := evmInterpreter.GetRequestMethod()
-	legacyRecord.ChainMethod = jsonrpcRequestMethod
-
-	// ErrorType is already set at gateway or protocol level.
-	// Skip updating the error fields to preserve the original error.
-	if legacyRecord.ErrorType != "" {
-		return legacyRecord
-	}
-
-	_, requestErr, err := evmInterpreter.GetRequestStatus()
-	// Could not extract request error details, skip the rest of the updates.
-	if err != nil || requestErr == nil {
-		return legacyRecord
-	}
-
-	legacyRecord.ErrorMessage = requestErr.String()
-
-	switch {
-	case requestErr.IsRequestError():
-		legacyRecord.ErrorType = fmt.Sprintf("%s_REQUEST_ERROR", qosEVMErrorTypeStr)
-	case requestErr.IsResponseError():
-		legacyRecord.ErrorType = fmt.Sprintf("%s_ENDPOINT_ERROR", qosEVMErrorTypeStr)
-	default:
-		legacyRecord.ErrorType = fmt.Sprintf("%s_UNKNOWN_ERROR", qosEVMErrorTypeStr)
-	}
-
-	return legacyRecord
-}
-
-func setLegacyFieldsFromMorseProtocolObservations(
-	logger polylog.Logger,
-	legacyRecord *legacyRecord,
-	observationList *protocolobservation.MorseObservationsList,
-) *legacyRecord {
-	// TODO_MVP(@adshmh): Simplify this if MorseObservationsList type is dropped in favor of using a single MorseRequestObservation per service request.
-	//
-	requestObservations := observationList.GetObservations()
-	if requestObservations == nil {
-		return legacyRecord
-	}
-	// Pick the last entry: this can be dropped once the above TODO is completed.
-	observations := requestObservations[len(requestObservations)-1]
-
-	// Use the ServiceID as the legacy record's chain ID.
-	legacyRecord.ChainID = observations.ServiceId
-
-	// Request processing error: set the fields and skip further processing.
-	if requestErr := observations.GetRequestError(); requestErr != nil {
-		legacyRecord.ErrorType = requestErr.ErrorType.String()
-		legacyRecord.ErrorMessage = requestErr.ErrorDetails
-
-		// Request error: no more data to add.
-		return legacyRecord
-	}
-
-	endpointObservations := observations.GetEndpointObservations()
-	// No endpoint observations: this should not happen as the request has not error set.
-	// Log a warning entry.
-	if len(endpointObservations) == 0 {
-		logger.Warn().Err(fmt.Errorf("")).Msg("Received no Morse endpoint observations for a valid request.")
-		return legacyRecord
-	}
-
-	// TODO_FUTURE(@adshmh): Update the processing method if a retry mechanism is implemented:
-	// Retries will result in multiple endpoint observations for a single request.
-	//
-	// Use the most recent entry in the endpoint observations.
-	endpointObservation := endpointObservations[len(endpointObservations)-1]
-
-	// Update error fields if an endpoint error has occurred.
-	legacyRecord = setLegacyErrFieldsFromMorseEndpointError(legacyRecord, endpointObservation)
-
-	legacyRecord.ProtocolAppPublicKey = endpointObservation.AppPublicKey
-
-	// Set endpoint query/response timestamps
-	legacyRecord.NodeQueryTimestamp = formatTimestampPbForBigQueryJSON(endpointObservation.EndpointQueryTimestamp)
-	legacyRecord.NodeReceiveTimestamp = formatTimestampPbForBigQueryJSON(endpointObservation.EndpointResponseTimestamp)
-
-	// track time spent waiting for the endpoint: required for calculating the `PortalTripTime` legacy field.
-	legacyRecord.endpointTripTime = endpointObservation.EndpointResponseTimestamp.AsTime().Sub(endpointObservation.EndpointQueryTimestamp.AsTime()).Seconds()
-
-	// Set endpoint address
-	legacyRecord.NodeAddress = endpointObservation.EndpointAddr
-
-	// Extract the endpoint's domain from its URL.
-	endpointDomain, err := extractEndpointDomainFromURL(endpointObservation.EndpointUrl)
-	// Error extracting the endpoint domain: log the error.
-	if err != nil {
-		logger.Warn().Err(fmt.Errorf("")).Msg("Received no Morse endpoint observations for a valid request.")
-	}
-	// Set the endpoint domain field: empty value if parsing the URL above failed.
-	legacyRecord.NodeDomain = endpointDomain
-
-	return legacyRecord
-}
-
-func setLegacyErrFieldsFromMorseEndpointError(
-	legacyRecord *legacyRecord,
-	endpointObservation *protocolobservation.MorseEndpointObservation,
-) *legacyRecord {
-
-	endpointErr := endpointObservation.ErrorType
-	// No endpoint error has occurred: no error processing required.
-	if endpointErr == nil {
-		return legacyRecord
-	}
-
-	// Update ErrorType using the observed endpoint error.
-	legacyRecord.ErrorType = endpointErr.String()
-
-	// Build the endpoint error details, including any sanctions.
-	var errMsg string
-	if errDetails := endpointObservation.GetErrorDetails(); errDetails != "" {
-		errMsg = fmt.Sprintf("error details: %s", errDetails)
-	}
-
-	// Add the sanction details to the error message.
-	if endpointSanction := endpointObservation.RecommendedSanction; endpointSanction != nil {
-		errMsg = fmt.Sprintf("%s, sanction: %s", errMsg, endpointSanction.String())
-	}
-
-	legacyRecord.ErrorMessage = errMsg
-
-	return legacyRecord
-}
-
-// formatTimestampPbForBigQueryJSON formats a protobuf Timestamp for BigQuery JSON inserts.
-// BigQuery expects timestamps in RFC 3339 format: YYYY-MM-DDTHH:MM:SS[.SSSSSS]Z
-func formatTimestampPbForBigQueryJSON(pbTimestamp *timestamppb.Timestamp) string {
-	// Convert the protobuf timestamp to Go time.Time
-	goTime := pbTimestamp.AsTime()
-
-	// Format in RFC 3339 format which BigQuery expects
-	return goTime.Format(time.RFC3339Nano)
-}
-
-// extractEndpointDomainFromURL extracts the domain name from a URL string.
-// It handles various URL formats including those with or without protocol,
-// subdomains, ports, paths, queries, and fragments.
-// Returns the domain without protocol, path, query parameters, or fragments.
-// Returns an empty string and error if the URL is invalid.
-func extractEndpointDomainFromURL(rawURL string) (string, error) {
-	// If the URL doesn't have a scheme, add one to make it parseable
-	if !strings.HasPrefix(rawURL, "http://") && !strings.HasPrefix(rawURL, "https://") {
-		rawURL = "https://" + rawURL
-	}
-
-	// Parse the URL
-	parsedURL, err := url.Parse(rawURL)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse URL: %w", err)
-	}
-
-	// Extract just the host part (domain + port if present)
-	host := parsedURL.Host
-
-	// Remove port number if present
-	if strings.Contains(host, ":") {
-		host = strings.Split(host, ":")[0]
-	}
-
-	return host, nil
 }
